tests: move tests now affecting mora.util to test_utils
