+++ backend/tests/test_service_common.py
@@ -6,9 +6,6 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 #
 
-import datetime
-
-import dateutil
 import freezegun
 
 from mora import common
@@ -27,76 +24,6 @@ class TestClass(util.TestCase):
             "500 Internal Server Error: Unknown Error.",
             str(exceptions.HTTPException()),
         )
-
-    def test_get_obj_path(self):
-        # Arrange
-        obj = {
-            'whatever': 'no',
-            'test1': {
-                'garbage': 'there is some stuff here already',
-                'test2': ['something']
-            }
-        }
-
-        path = ('test1', 'test2')
-
-        expected_props = ['something']
-
-        # Act
-        actual_props = mora_util.get_obj_value(obj, path)
-
-        # Assert
-        self.assertEqual(expected_props, actual_props)
-
-    def test_get_obj_path_none(self):
-        # Arrange
-        obj = {
-            'whatever': 'no',
-            'test1': None,
-        }
-
-        path = ('test1', 'test2')
-
-        expected_props = None
-
-        # Act
-        actual_props = mora_util.get_obj_value(obj, path)
-
-        # Assert
-        self.assertEqual(expected_props, actual_props)
-
-    def test_get_obj_path_missing(self):
-        # Arrange
-        obj = {
-            'whatever': 'no',
-        }
-
-        path = ('test1',)
-
-        expected_props = None
-
-        # Act
-        actual_props = mora_util.get_obj_value(obj, path)
-
-        # Assert
-        self.assertEqual(expected_props, actual_props)
-
-    def test_get_obj_path_weird(self):
-        # Arrange
-        obj = {
-            'whatever': 'no',
-            'test1': 42,
-        }
-
-        path = ('test1', 'test2')
-
-        expected_props = None
-
-        # Act
-        actual_props = mora_util.get_obj_value(obj, path)
-
-        # Assert
-        self.assertEqual(expected_props, actual_props)
 
     def test_update_payload_complex(self):
         # Arrange
@@ -1881,460 +1808,6 @@ class TestClass(util.TestCase):
 
         # Assert
         self.assertEqual(expected_result, actual_result)
-
-    def test_set_obj_value_existing_path(self):
-        # Arrange
-        obj = {'test1': {'test2': [{'key1': 'val1'}]}}
-        path = ('test1', 'test2')
-
-        val = [{'key2': 'val2'}]
-
-        expected_result = {
-            'test1': {
-                'test2': [
-                    {'key1': 'val1'},
-                    {'key2': 'val2'},
-                ]
-            }
-        }
-
-        # Act
-        actual_result = mora_util.set_obj_value(obj, path, val)
-
-        # Assert
-        self.assertEqual(expected_result, actual_result)
-
-    def test_set_obj_value_new_path(self):
-        # Arrange
-        obj = {}
-        path = ('test1', 'test2')
-
-        val = [{'key2': 'val2'}]
-
-        expected_result = {
-            'test1': {
-                'test2': [
-                    {'key2': 'val2'},
-                ]
-            }
-        }
-
-        # Act
-        actual_result = mora_util.set_obj_value(obj, path, val)
-
-        # Assert
-        self.assertEqual(expected_result, actual_result)
-
-    def test_get_valid_from(self):
-        ts = '2018-03-21T00:00:00+01:00'
-        dt = datetime.datetime(2018, 3, 21,
-                               tzinfo=dateutil.tz.tzoffset(None, 3600))
-
-        self.assertEqual(dt, mora_util.get_valid_from(
-            {
-                'validity': {
-                    'from': ts,
-                }
-            },
-        ))
-
-        self.assertEqual(dt, mora_util.get_valid_from(
-            {
-                'validity': {
-                },
-            },
-            {
-                'validity': {
-                    'from': ts,
-                }
-            }
-        ))
-
-        self.assertRaises(
-            exceptions.HTTPException, mora_util.get_valid_from,
-            {},
-        )
-
-        self.assertRaises(
-            exceptions.HTTPException, mora_util.get_valid_from,
-            {
-                'validity': {},
-            },
-        )
-
-        self.assertRaises(
-            exceptions.HTTPException, mora_util.get_valid_from,
-            {},
-            {
-                'validity': {
-                }
-            },
-        )
-
-        self.assertRaises(
-            exceptions.HTTPException, mora_util.get_valid_from,
-            {
-
-            },
-            {
-                'validity': {
-                }
-            },
-        )
-
-        self.assertRaises(
-            exceptions.HTTPException, mora_util.get_valid_from,
-            {
-
-            },
-            {
-                'validity': {
-                    'from': None,
-                }
-            },
-        )
-
-    def test_get_valid_to(self):
-        ts = '2018-03-21T00:00:00+01:00'
-        dt = datetime.datetime(2018, 3, 21,
-                               tzinfo=dateutil.tz.tzoffset(None, 3600))
-
-        self.assertEqual(dt, mora_util.get_valid_to(
-            {
-                'validity': {
-                    'to': ts,
-                }
-            },
-        ))
-
-        self.assertEqual(dt, mora_util.get_valid_to(
-            {
-                'validity': {
-                },
-            },
-            {
-                'validity': {
-                    'to': ts,
-                }
-            }
-        ))
-
-        self.assertEqual(
-            mora_util.POSITIVE_INFINITY,
-            mora_util.get_valid_to({}),
-        )
-
-        self.assertEqual(
-            mora_util.get_valid_to({
-                'validity': {},
-            }),
-            mora_util.POSITIVE_INFINITY,
-        )
-
-        self.assertEqual(
-            mora_util.POSITIVE_INFINITY,
-            mora_util.get_valid_to(
-                {},
-                {
-                    'validity': {
-                    }
-                },
-            ),
-        )
-
-        self.assertEqual(
-            mora_util.POSITIVE_INFINITY,
-            mora_util.get_valid_to(
-                {
-                    'validity': {
-                        'to': None,
-                    }
-                },
-            ),
-        )
-
-        self.assertEqual(
-            mora_util.POSITIVE_INFINITY,
-            mora_util.get_valid_to(
-                {},
-                {
-                    'validity': {
-                        'to': None,
-                    }
-                },
-            ),
-        )
-
-    def test_get_validities(self):
-        # start time required
-        self.assertRaises(
-            exceptions.HTTPException,
-            mora_util.get_valid_from, {}, {},
-        )
-
-        self.assertRaises(
-            exceptions.HTTPException,
-            mora_util.get_valid_from, {}, {
-                'validity': None,
-            },
-        )
-
-        self.assertRaises(
-            exceptions.HTTPException,
-            mora_util.get_valid_from, {}, {
-                'validity': {
-                    'from': None,
-                },
-            },
-        )
-
-        # still nothing
-        self.assertEqual(
-            mora_util.get_valid_to({}, {}),
-            mora_util.POSITIVE_INFINITY,
-        )
-
-        self.assertEqual(
-            mora_util.get_valid_to({}, {
-                'validity': None,
-            }),
-            mora_util.POSITIVE_INFINITY,
-        )
-
-        self.assertEqual(
-            mora_util.POSITIVE_INFINITY,
-            mora_util.get_valid_to({}, {
-                'validity': {
-                    'to': None,
-                },
-            }),
-        )
-
-        # actually set
-        self.assertEqual(
-            datetime.datetime(2018, 3, 5, tzinfo=mora_util.DEFAULT_TIMEZONE),
-            mora_util.get_valid_from({
-                'validity': {
-                    'from': '2018-03-05',
-                },
-            }),
-        )
-
-        self.assertEqual(
-            datetime.datetime(2018, 3, 5, tzinfo=mora_util.DEFAULT_TIMEZONE),
-            mora_util.get_valid_to({
-                'validity': {
-                    'to': '2018-03-05',
-                },
-            }),
-        )
-
-        # actually set in the fallback
-        self.assertEqual(
-            datetime.datetime(2018, 3, 5, tzinfo=mora_util.DEFAULT_TIMEZONE),
-            mora_util.get_valid_from({}, {
-                'validity': {
-                    'from': '2018-03-05',
-                },
-            }),
-        )
-
-        self.assertEqual(
-            datetime.datetime(2018, 3, 5, tzinfo=mora_util.DEFAULT_TIMEZONE),
-            mora_util.get_valid_to({}, {
-                'validity': {
-                    'to': '2018-03-05',
-                },
-            }),
-        )
-
-        self.assertEqual(
-            (datetime.datetime(2018, 3, 5, tzinfo=mora_util.DEFAULT_TIMEZONE),
-             datetime.datetime(2018, 4, 5, tzinfo=mora_util.DEFAULT_TIMEZONE)),
-            mora_util.get_validities({
-                'validity': {
-                    'from': '2018-03-05',
-                    'to': '2018-04-05',
-                },
-            }),
-        )
-
-        self.assertEqual(
-            (datetime.datetime(2018, 3, 5, tzinfo=mora_util.DEFAULT_TIMEZONE),
-             mora_util.POSITIVE_INFINITY),
-            mora_util.get_validities({
-                'validity': {
-                    'from': '2018-03-05'
-                },
-            }),
-        )
-
-        with self.assertRaisesRegex(exceptions.HTTPException,
-                                    "End date is before start date"):
-            mora_util.get_validities({
-                'validity': {
-                    'from': '2019-03-05',
-                    'to': '2018-03-05',
-                },
-            })
-
-    def test_get_uuid(self):
-        testid = '00000000-0000-0000-0000-000000000000'
-
-        self.assertEqual(
-            testid,
-            mora_util.get_uuid({
-                'uuid': testid,
-            }),
-        )
-
-        self.assertEqual(
-            testid,
-            mora_util.get_uuid(
-                {},
-                {
-                    'uuid': testid,
-                },
-            ),
-        )
-
-        self.assertRaises(
-            exceptions.HTTPException,
-            mora_util.get_uuid,
-            {
-                'uuid': 42,
-            },
-        )
-
-        self.assertEqual(
-            None,
-            mora_util.get_uuid(
-                {},
-                required=False,
-            ),
-        )
-
-        self.assertEqual(
-            testid,
-            mora_util.get_uuid(
-                {
-                    'kaflaflibob': testid,
-                    'uuid': 42,
-                },
-                key='kaflaflibob',
-            ),
-        )
-
-    def test_checked_get(self):
-        mapping = {
-            'list': [1337],
-            'dict': {1337: 1337},
-            'string': '1337',
-            'int': 1337,
-            'null': None,
-        }
-
-        # when it's there
-        self.assertIs(
-            mora_util.checked_get(mapping, 'list', []),
-            mapping['list'],
-        )
-
-        self.assertIs(
-            mora_util.checked_get(mapping, 'dict', {}),
-            mapping['dict'],
-        )
-
-        self.assertIs(
-            mora_util.checked_get(mapping, 'string', ''),
-            mapping['string'],
-        )
-
-        self.assertIs(
-            mora_util.checked_get(mapping, 'int', 1337),
-            mapping['int'],
-        )
-
-        # when it's not there
-        self.assertEqual(
-            mora_util.checked_get(mapping, 'nonexistent', []),
-            [],
-        )
-
-        self.assertEqual(
-            mora_util.checked_get(mapping, 'nonexistent', {}),
-            {},
-        )
-
-        self.assertEqual(
-            mora_util.checked_get(mapping, 'null', {}),
-            {},
-        )
-
-        with self.assertRaisesRegex(exceptions.HTTPException,
-                                    "Missing nonexistent"):
-            mora_util.checked_get(mapping, 'nonexistent', [], required=True)
-
-        with self.assertRaisesRegex(exceptions.HTTPException,
-                                    "Missing nonexistent"):
-            mora_util.checked_get(mapping, 'nonexistent', {}, required=True)
-
-        # bad value
-        with self.assertRaisesRegex(
-                exceptions.HTTPException,
-                'Invalid \'dict\', expected list, got: {"1337": 1337}',
-        ):
-            mora_util.checked_get(mapping, 'dict', [])
-
-        with self.assertRaisesRegex(
-                exceptions.HTTPException,
-                r"Invalid 'list', expected dict, got: \[1337\]",
-        ):
-            mora_util.checked_get(mapping, 'list', {})
-
-    def test_get_urn(self):
-        with self.subTest('bad string'):
-            with self.assertRaisesRegex(
-                exceptions.HTTPException,
-                "invalid urn for 'urn': '42'",
-            ) as ctxt:
-                mora_util.get_urn({'urn': '42'})
-
-            self.assertEqual(
-                {
-                    'description': "invalid urn for 'urn': '42'",
-                    'error': True,
-                    'error_key': 'E_INVALID_URN',
-                    'obj': {'urn': '42'},
-                    'status': 400,
-                },
-                ctxt.exception.response.json,
-            )
-
-            self.assertEqual(
-                "400 Bad Request: invalid urn for 'urn': '42'",
-                str(ctxt.exception),
-            )
-
-        with self.assertRaisesRegex(
-            exceptions.HTTPException,
-            "Invalid 'urn', expected str, got: 42",
-        ) as ctxt:
-            mora_util.get_urn({'urn': 42})
-
-        self.assertEqual(
-            {
-                'description': "Invalid 'urn', expected str, got: 42",
-                'error': True,
-                'error_key': 'E_INVALID_TYPE',
-                'expected': 'str',
-                'actual': '42',
-                'key': 'urn',
-                'obj': {'urn': 42},
-                'status': 400,
-            },
-            ctxt.exception.response.json,
-        )
 
     @freezegun.freeze_time('2018-01-01')
     @util.mock()
